+++ config/checkstyle/checkstyle.xml
@@ -90,7 +90,6 @@
     <module name="ModifierOrder"/>
     <module name="EmptyLineSeparator">
       <property name="allowNoEmptyLineBetweenFields" value="true"/>
-      <property name="tokens" value="VARIABLE_DEF, METHOD_DEF"/>
     </module>
     <module name="SeparatorWrap">
       <property name="id" value="SeparatorWrapDot"/>
