+++ config/pmd/ruleset.xml
@@ -21,11 +21,6 @@
     <exclude name="DefaultPackage"/>
     <exclude name="CommentDefaultAccessModifier"/>
   </rule>
-  <rule ref="category/java/design.xml/EmptyLineSeparator">
-    <properties>
-      <property name="tokens" value="VARIABLE_DEF, METHOD_DEF"/>
-    </properties>
-  </rule>
   <rule ref="category/java/codestyle.xml/ClassNamingConventions">
     <properties>
       <!-- same as any other class -->
