+++ src/main/java/uk/gov/hmcts/reform/prl/models/Organisation.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.extern.jackson.Jacksonized;
 
+import java.util.UUID;
+
 @Data
 @Builder
 @Jacksonized
@@ -15,6 +17,17 @@ public class Organisation {
     private String organisationID;
     @JsonProperty("OrganisationName")
     private String organisationName;
+
+    private UUID solicitorOrgId;
+    public void setSolicitorOrgId(UUID solicitorOrgId) {
+        if (this.getSolicitorOrgId() == null) {
+            if (solicitorOrgId != null) {
+                this.solicitorOrgId = solicitorOrgId;
+            } else {
+                this.solicitorOrgId = UUID.randomUUID();
+            }
+        }
+    }
 
     public static Organisation organisation(String id) {
         return Organisation.builder()
